requirements.txt
python-telegram-bot==13.15
redis==5.0.1
python-dotenv==1.0.1
vk_api==11.10.0